Add `where` parameter to safeguards checks and safe intervals, reimplement `monotonicity` with QoIs, and add boolean QoI functions (#48)
* Add safeguard where for partial evaluation

* Fix finite difference denominator precedence

* Implement compute_footprint and compute_inverse_footprint for pointwise and stencil safeguards

* Ignore buggy numpy_quaddtype invalid cast warnings

* Simplify docs with relative crossrefs

* Fix QoI expr args for >=2d array

* Support literal index for select safeguard

* Add isosurface example

* Add correction metrics to examples

* Typing improvements

* Add regression tests for shadowed stencil requirements

* Improve select.selector fuzzing

* Implement not, all, any QoI combinators

* Support slice indexing in QoIs

* Add <,<=,>=,> in QoIs

* Fix np.place for quaddtype

* Improve QoI divmul const zero bounds

* Upgrade to SZ3 v3.3

* Find sparks of joy in example plotting

* Add xfail test for https://github.com/numpy/numpy-user-dtypes/issues/239

* Implement iterative corrections using where

* Implement more flexible QoI centre and shape function

* Update the QoI grammar

+++ bench/compression.py
@@ -138,7 +138,7 @@ class MyLinearQuantizer(MyQuantizer):
         self._eb_abs = eb_abs
 
     def encoded_dtype(self, dtype: np.dtype) -> np.dtype:
-        return np.dtype(int)
+        return np.dtype(np.intp)
 
     def encode(self, x, predict):
         return np.round((x - predict) / (self._eb_abs * 2))
